Fully implemented but incorrect

+++ Numerical Integration.h
@@ -18,6 +18,6 @@
 arma::vec grid_int_vec(const std::function<double(double,double,double)> &f, int Ng, double L);
 
 // More brute force
-double quad_3D_grid_integration(const std::function<double(double,double,double)> &f, double Ng, double L);
+double quad_3D_grid_integration(const std::function<double(double,double,double)> &f, int Ng, double L);
 
 #endif //HW6_NUMERICAL_INTEGRATION_H
